fix: coupling SBR issue (ELD only)
When in SWITCH_LRC mode, the encoder attempts to make a choice between LR and COUPLING SBR envelop encoding.
Both LR and COUPLING envelops are calculated and quantized correctly.
If COUPLING decision is made the bitstream formatter uses 'channel specific' compression tables for RIGHT channel to encode the envelop, but decoder expects same tables as for the LEFT.
Actually, encoder uses wrong tables already at a bit-budget calculation stage.

Existing fix prevent using SWITCH_LRC mode if compression tables for LEFT and RIGHT channels are different. This is defined individually for each frame based on source material if tables are different or same. At the begging, the encoder estimates the precision necessary to quantize tone and noise envelops, it could be 1.5 or 3dB. This defines which compression tables to use.

> fdkaac.exe -p 39 -b 24000 -o out.mp4 -f 0 --lowdelay-sbr -1 "BlackBird.wav"
> m4a-dec.exe out.mp4 out.wav

Existing solution is not good enough. It might be much better to fix tables used in compression procedures. Unfortunately, tables for bit-budget evaluation and stream writing (both wrong) comes from different structures in a c-code and it make take a long time to fix up the problem without breaking other dependencies between these data structures and surrounding code.